Update and changes in the Configuration page in the docs
I've added the option and table entry for the configurable archive grouping
mentioned in issue #8. Fixes #8.

I've also changed the order in which the options and table entries related
to previews were placed. Its now ordered like:

* previews?
* posts-per-page
* blocks-per-preview

This is so that a new user will read about the preview? option first before
reading about post-per-page, which first requires knowledge on what
post previews are (I did this because I got confused the first time I
saw it the old way :D).